+++ CoR/remap_grid_data_class.h
@@ -54,8 +54,8 @@ class Remap_grid_data_class
         void generate_analytic_values(const char*);
         void evaluate_error(Remap_grid_data_class*, Remap_grid_data_class*);
         void change_datatype_in_application(const char*);
-		void write_grid_data_into_array(char **, int &, int &);
-		Remap_grid_data_class(Remap_grid_class *, const char *, int&);
+		void write_grid_data_into_array(char **, long &, long &);
+		Remap_grid_data_class(Remap_grid_class *, const char *, long&);
 };
